Add a new benchmark for proof generation
